Phase B8: extract shared helpers from main.rs into dedicated modules
Move 1600+ lines out of main.rs into:
- graphql_access.rs: post_graphql, graphql_rows, field helpers, etc.
- request_helpers.rs: SubmittedRequest, create_agent_request, wait_for_terminal_response, etc.
- config_bundle.rs: build_config_export_bundle, build_desired_state_live_bundle, sanitize_import_document
- config_import.rs: read/validate/apply import bundle functions
- home_state.rs: resolve_home_dir, read/write init config and runtime state
- resolve_helpers.rs: backend resolution, task prompt, tool mode normalization

main.rs is now 573 lines (398 non-test code).
