+++ crates/defra-agent-cli/src/config_bundle.rs
@@ -1,0 +1,534 @@
+use std::collections::BTreeSet;
+
+use anyhow::Result;
+use defra_agent::graphql::escape_graphql_string;
+use serde_json::Value;
+
+use crate::desired_state;
+use crate::shared::ConfigExportBundle;
+use crate::{
+    graphql_rows, graphql_rows_or_empty_if_collection_missing, graphql_string_list_literal,
+    EXPORT_AGENT_BEHAVIOR_FIELDS, EXPORT_AGENT_PRINCIPAL_FIELDS, EXPORT_INFERENCE_BACKEND_FIELDS,
+    EXPORT_INFERENCE_PROFILE_FIELDS, EXPORT_SCHEDULED_TASK_FIELDS, EXPORT_TOOL_SELECTION_FIELDS,
+    EXPORT_TOOL_SERVICE_REGISTRY_FIELDS, CONFIG_EXPORT_FORMAT,
+};
+use crate::config_writes::ConfigAccess;
+
+pub(crate) async fn build_config_export_bundle(
+    access: &ConfigAccess,
+    agent_did: &str,
+) -> Result<ConfigExportBundle> {
+    let principal_rows = graphql_rows(
+        access,
+        "AgentPrincipal",
+        &format!(
+            r#"{{
+                AgentPrincipal(
+                    filter: {{ agent_did: {{ _eq: "{agent_did}" }} }},
+                    limit: 1
+                ) {{
+                    {fields}
+                }}
+            }}"#,
+            agent_did = escape_graphql_string(agent_did),
+            fields = EXPORT_AGENT_PRINCIPAL_FIELDS,
+        ),
+    )
+    .await?;
+    let mut behavior_rows = graphql_rows(
+        access,
+        "AgentBehavior",
+        &format!(
+            r#"{{
+                AgentBehavior(
+                    filter: {{ agent_did: {{ _eq: "{agent_did}" }} }},
+                    order: {{ created_at: ASC }}
+                ) {{
+                    {fields}
+                }}
+            }}"#,
+            agent_did = escape_graphql_string(agent_did),
+            fields = EXPORT_AGENT_BEHAVIOR_FIELDS,
+        ),
+    )
+    .await?;
+    sort_document_rows(&mut behavior_rows, "behavior_id");
+
+    let tool_selection_ids = collect_string_field_values(&behavior_rows, "tool_selection_id");
+    let backend_ids = collect_string_field_values(&behavior_rows, "backend_id");
+    let profile_ids = collect_string_field_values(&behavior_rows, "inference_profile_id");
+
+    let mut tool_selection_rows = if tool_selection_ids.is_empty() {
+        Vec::new()
+    } else {
+        graphql_rows(
+            access,
+            "ToolSelection",
+            &format!(
+                r#"{{
+                    ToolSelection(
+                        filter: {{ selection_id: {{ _in: {} }} }}
+                    ) {{
+                        {fields}
+                    }}
+                }}"#,
+                graphql_string_list_literal(&tool_selection_ids),
+                fields = EXPORT_TOOL_SELECTION_FIELDS,
+            ),
+        )
+        .await?
+    };
+    sort_document_rows(&mut tool_selection_rows, "selection_id");
+
+    let mut backend_rows = if backend_ids.is_empty() {
+        Vec::new()
+    } else {
+        graphql_rows(
+            access,
+            "InferenceBackend",
+            &format!(
+                r#"{{
+                    InferenceBackend(
+                        filter: {{ backend_id: {{ _in: {} }} }}
+                    ) {{
+                        {fields}
+                    }}
+                }}"#,
+                graphql_string_list_literal(&backend_ids),
+                fields = EXPORT_INFERENCE_BACKEND_FIELDS,
+            ),
+        )
+        .await?
+    };
+    sort_document_rows(&mut backend_rows, "backend_id");
+
+    let mut profile_rows = if profile_ids.is_empty() {
+        Vec::new()
+    } else {
+        graphql_rows(
+            access,
+            "InferenceProfile",
+            &format!(
+                r#"{{
+                    InferenceProfile(
+                        filter: {{ profile_id: {{ _in: {} }} }}
+                    ) {{
+                        {fields}
+                    }}
+                }}"#,
+                graphql_string_list_literal(&profile_ids),
+                fields = EXPORT_INFERENCE_PROFILE_FIELDS,
+            ),
+        )
+        .await?
+    };
+    sort_document_rows(&mut profile_rows, "profile_id");
+    let mut tool_service_registry_rows = graphql_rows_or_empty_if_collection_missing(
+        access,
+        "ToolServiceRegistry",
+        &format!(
+            r#"{{
+                ToolServiceRegistry {{
+                    {fields}
+                }}
+            }}"#,
+            fields = EXPORT_TOOL_SERVICE_REGISTRY_FIELDS,
+        ),
+    )
+    .await?;
+    sort_document_rows(&mut tool_service_registry_rows, "service_id");
+    normalize_tool_service_registry_export_rows(&mut tool_service_registry_rows)?;
+    let mut scheduled_task_rows = graphql_rows_or_empty_if_collection_missing(
+        access,
+        "ScheduledTask",
+        &format!(
+            r#"{{
+                ScheduledTask(
+                    filter: {{ agent_did: {{ _eq: "{agent_did}" }} }}
+                ) {{
+                    {fields}
+                }}
+            }}"#,
+            agent_did = escape_graphql_string(agent_did),
+            fields = EXPORT_SCHEDULED_TASK_FIELDS,
+        ),
+    )
+    .await?;
+    sort_document_rows(&mut scheduled_task_rows, "task_id");
+
+    Ok(ConfigExportBundle {
+        format: CONFIG_EXPORT_FORMAT.to_string(),
+        agent_did: agent_did.to_string(),
+        exported_at: chrono::Utc::now().to_rfc3339(),
+        access_mode: access.mode().to_string(),
+        agent_principal: principal_rows.into_iter().next(),
+        agent_behaviors: behavior_rows,
+        tool_selections: tool_selection_rows,
+        inference_backends: backend_rows,
+        inference_profiles: profile_rows,
+        tool_service_registries: tool_service_registry_rows,
+        scheduled_tasks: scheduled_task_rows,
+    })
+}
+
+pub(crate) async fn build_desired_state_live_bundle(
+    access: &ConfigAccess,
+    desired_manifest: &desired_state::DesiredStateManifest,
+) -> Result<ConfigExportBundle> {
+    let agent_did = desired_manifest.agent_principal.agent_did.as_str();
+    let principal_rows = graphql_rows(
+        access,
+        "AgentPrincipal",
+        &format!(
+            r#"{{
+                AgentPrincipal(
+                    filter: {{ agent_did: {{ _eq: "{agent_did}" }} }},
+                    limit: 1
+                ) {{
+                    {fields}
+                }}
+            }}"#,
+            agent_did = escape_graphql_string(agent_did),
+            fields = EXPORT_AGENT_PRINCIPAL_FIELDS,
+        ),
+    )
+    .await?;
+    let mut behavior_rows = graphql_rows(
+        access,
+        "AgentBehavior",
+        &format!(
+            r#"{{
+                AgentBehavior(
+                    filter: {{ agent_did: {{ _eq: "{agent_did}" }} }},
+                    order: {{ created_at: ASC }}
+                ) {{
+                    {fields}
+                }}
+            }}"#,
+            agent_did = escape_graphql_string(agent_did),
+            fields = EXPORT_AGENT_BEHAVIOR_FIELDS,
+        ),
+    )
+    .await?;
+    sort_document_rows(&mut behavior_rows, "behavior_id");
+
+    let tool_selection_ids = collect_string_field_values(&behavior_rows, "tool_selection_id")
+        .into_iter()
+        .chain(
+            desired_manifest
+                .tool_selections
+                .iter()
+                .map(|value| value.selection_id.clone()),
+        )
+        .collect::<BTreeSet<_>>()
+        .into_iter()
+        .collect::<Vec<_>>();
+    let backend_ids = collect_string_field_values(&behavior_rows, "backend_id")
+        .into_iter()
+        .chain(
+            desired_manifest
+                .inference_backends
+                .iter()
+                .map(|value| value.backend_id.clone()),
+        )
+        .collect::<BTreeSet<_>>()
+        .into_iter()
+        .collect::<Vec<_>>();
+    let profile_ids = collect_string_field_values(&behavior_rows, "inference_profile_id")
+        .into_iter()
+        .chain(
+            desired_manifest
+                .inference_profiles
+                .iter()
+                .map(|value| value.profile_id.clone()),
+        )
+        .collect::<BTreeSet<_>>()
+        .into_iter()
+        .collect::<Vec<_>>();
+
+    let mut tool_selection_rows = if tool_selection_ids.is_empty() {
+        Vec::new()
+    } else {
+        graphql_rows(
+            access,
+            "ToolSelection",
+            &format!(
+                r#"{{
+                    ToolSelection(
+                        filter: {{ selection_id: {{ _in: {} }} }}
+                    ) {{
+                        {fields}
+                    }}
+                }}"#,
+                graphql_string_list_literal(&tool_selection_ids),
+                fields = EXPORT_TOOL_SELECTION_FIELDS,
+            ),
+        )
+        .await?
+    };
+    sort_document_rows(&mut tool_selection_rows, "selection_id");
+
+    let mut backend_rows = if backend_ids.is_empty() {
+        Vec::new()
+    } else {
+        graphql_rows(
+            access,
+            "InferenceBackend",
+            &format!(
+                r#"{{
+                    InferenceBackend(
+                        filter: {{ backend_id: {{ _in: {} }} }}
+                    ) {{
+                        {fields}
+                    }}
+                }}"#,
+                graphql_string_list_literal(&backend_ids),
+                fields = EXPORT_INFERENCE_BACKEND_FIELDS,
+            ),
+        )
+        .await?
+    };
+    sort_document_rows(&mut backend_rows, "backend_id");
+
+    let mut profile_rows = if profile_ids.is_empty() {
+        Vec::new()
+    } else {
+        graphql_rows(
+            access,
+            "InferenceProfile",
+            &format!(
+                r#"{{
+                    InferenceProfile(
+                        filter: {{ profile_id: {{ _in: {} }} }}
+                    ) {{
+                        {fields}
+                    }}
+                }}"#,
+                graphql_string_list_literal(&profile_ids),
+                fields = EXPORT_INFERENCE_PROFILE_FIELDS,
+            ),
+        )
+        .await?
+    };
+    sort_document_rows(&mut profile_rows, "profile_id");
+    let tool_service_ids = desired_manifest
+        .tool_service_registries
+        .iter()
+        .map(|value| value.service_id.clone())
+        .collect::<BTreeSet<_>>()
+        .into_iter()
+        .collect::<Vec<_>>();
+    let mut tool_service_registry_rows = if tool_service_ids.is_empty() {
+        Vec::new()
+    } else {
+        graphql_rows_or_empty_if_collection_missing(
+            access,
+            "ToolServiceRegistry",
+            &format!(
+                r#"{{
+                    ToolServiceRegistry(
+                        filter: {{ service_id: {{ _in: {} }} }}
+                    ) {{
+                        {fields}
+                    }}
+                }}"#,
+                graphql_string_list_literal(&tool_service_ids),
+                fields = EXPORT_TOOL_SERVICE_REGISTRY_FIELDS,
+            ),
+        )
+        .await?
+    };
+    sort_document_rows(&mut tool_service_registry_rows, "service_id");
+    let mut scheduled_task_rows = graphql_rows_or_empty_if_collection_missing(
+        access,
+        "ScheduledTask",
+        &format!(
+            r#"{{
+                ScheduledTask(
+                    filter: {{ agent_did: {{ _eq: "{agent_did}" }} }}
+                ) {{
+                    {fields}
+                }}
+            }}"#,
+            agent_did = escape_graphql_string(agent_did),
+            fields = EXPORT_SCHEDULED_TASK_FIELDS,
+        ),
+    )
+    .await?;
+    sort_document_rows(&mut scheduled_task_rows, "task_id");
+
+    Ok(ConfigExportBundle {
+        format: CONFIG_EXPORT_FORMAT.to_string(),
+        agent_did: agent_did.to_string(),
+        exported_at: chrono::Utc::now().to_rfc3339(),
+        access_mode: access.mode().to_string(),
+        agent_principal: principal_rows.into_iter().next(),
+        agent_behaviors: behavior_rows,
+        tool_selections: tool_selection_rows,
+        inference_backends: backend_rows,
+        inference_profiles: profile_rows,
+        tool_service_registries: tool_service_registry_rows,
+        scheduled_tasks: scheduled_task_rows,
+    })
+}
+
+pub(crate) fn live_manifest_from_bundle(
+    desired_manifest: &desired_state::DesiredStateManifest,
+    live_bundle: &ConfigExportBundle,
+) -> Result<(
+    Option<desired_state::DesiredAgentPrincipal>,
+    desired_state::DesiredStateManifest,
+)> {
+    if live_bundle.agent_principal.is_some() {
+        let live_manifest = desired_state::manifest_from_export_bundle(live_bundle)?;
+        Ok((Some(live_manifest.agent_principal.clone()), live_manifest))
+    } else {
+        Ok((
+            None,
+            desired_state::DesiredStateManifest {
+                agent_principal: desired_manifest.agent_principal.clone(),
+                agent_behaviors: Vec::new(),
+                tool_selections: Vec::new(),
+                inference_backends: Vec::new(),
+                inference_profiles: Vec::new(),
+                tool_service_registries: Vec::new(),
+                scheduled_tasks: Vec::new(),
+            },
+        ))
+    }
+}
+
+pub(crate) fn sort_document_rows(rows: &mut [Value], key: &str) {
+    rows.sort_by(|left, right| {
+        let left_key = left.get(key).and_then(Value::as_str).unwrap_or_default();
+        let right_key = right.get(key).and_then(Value::as_str).unwrap_or_default();
+        left_key.cmp(right_key)
+    });
+}
+
+pub(crate) fn normalize_tool_service_registry_export_rows(rows: &mut [Value]) -> Result<()> {
+    for row in rows {
+        let object = row
+            .as_object_mut()
+            .ok_or_else(|| anyhow::anyhow!("ToolServiceRegistry export row must be an object"))?;
+        desired_state::normalize_tool_service_registry_storage_fields(object)?;
+    }
+    Ok(())
+}
+
+pub(crate) fn collect_string_field_values(rows: &[Value], field: &str) -> Vec<String> {
+    let mut values = rows
+        .iter()
+        .filter_map(|row| row.get(field).and_then(Value::as_str))
+        .map(str::trim)
+        .filter(|value| !value.is_empty())
+        .map(ToOwned::to_owned)
+        .collect::<Vec<_>>();
+    values.sort();
+    values.dedup();
+    values
+}
+
+pub(crate) fn sanitize_import_document(collection_name: &str, doc: &Value, for_update: bool) -> Result<Value> {
+    let mut object = match collection_name {
+        "InferenceBackend" | "ScheduledTask" | "ToolServiceRegistry" => {
+            doc.as_object().cloned().ok_or_else(|| {
+                anyhow::anyhow!("{collection_name} import document must be an object")
+            })?
+        }
+        _ => return Ok(doc.clone()),
+    };
+
+    match collection_name {
+        "InferenceBackend" => {
+            desired_state::strip_deprecated_inference_backend_fields(&mut object);
+            object.remove("last_probe");
+            if for_update {
+                object.insert("last_probe".to_string(), Value::Null);
+            }
+        }
+        "ScheduledTask" => {
+            for field in [
+                "next_run_at",
+                "last_run_at",
+                "last_status",
+                "last_error",
+                "run_count",
+                "created_at",
+                "updated_at",
+            ] {
+                object.remove(field);
+            }
+            if for_update {
+                object.insert("next_run_at".to_string(), Value::Null);
+                object.insert("last_run_at".to_string(), Value::Null);
+                object.insert("created_at".to_string(), Value::Null);
+                object.insert("updated_at".to_string(), Value::Null);
+            }
+        }
+        "ToolServiceRegistry" => {
+            for field in ["tools", "version", "updated_at"] {
+                object.remove(field);
+            }
+            desired_state::normalize_tool_service_registry_storage_fields(&mut object)?;
+            if for_update {
+                object.insert("updated_at".to_string(), Value::Null);
+            }
+            match object.get("status") {
+                Some(Value::String(s)) if !s.is_empty() => {}
+                _ => {
+                    object.insert("status".to_string(), Value::String("online".to_string()));
+                }
+            }
+        }
+        _ => unreachable!(),
+    }
+
+    Ok(Value::Object(object))
+}
+
+pub(crate) fn select_apply_collection_docs(
+    docs: &[Value],
+    unique_field: &str,
+    collection_name: &str,
+    diff: &desired_state::DesiredStateCollectionDiff,
+) -> Result<Vec<Value>> {
+    let requested_ids = diff
+        .create
+        .iter()
+        .chain(diff.update.iter())
+        .cloned()
+        .collect::<BTreeSet<_>>();
+    if requested_ids.is_empty() {
+        return Ok(Vec::new());
+    }
+
+    let mut selected = docs
+        .iter()
+        .filter(|doc| {
+            doc.get(unique_field)
+                .and_then(Value::as_str)
+                .is_some_and(|value| requested_ids.contains(value))
+        })
+        .cloned()
+        .collect::<Vec<_>>();
+    sort_document_rows(&mut selected, unique_field);
+
+    let found_ids = selected
+        .iter()
+        .filter_map(|doc| doc.get(unique_field).and_then(Value::as_str))
+        .map(ToOwned::to_owned)
+        .collect::<BTreeSet<_>>();
+    let missing_ids = requested_ids
+        .difference(&found_ids)
+        .cloned()
+        .collect::<Vec<_>>();
+    if !missing_ids.is_empty() {
+        anyhow::bail!(
+            "desired-state apply missing {collection_name} documents for ids: {}",
+            missing_ids.join(", ")
+        );
+    }
+
+    Ok(selected)
+}
