+++ crates/defra-agent-cli/src/graphql_access.rs
@@ -1,0 +1,236 @@
+use anyhow::{Context, Result};
+use defra_agent::graphql::escape_graphql_string;
+use serde_json::{json, Value};
+
+use crate::config_writes::ConfigAccess;
+
+pub(crate) async fn graphql_endpoint_available(graphql: &str) -> bool {
+    let client = match reqwest::Client::builder()
+        .timeout(std::time::Duration::from_secs(2))
+        .build()
+    {
+        Ok(client) => client,
+        Err(_) => return false,
+    };
+    match client
+        .post(graphql)
+        .json(&json!({ "query": "{ __typename }" }))
+        .send()
+        .await
+    {
+        Ok(response) => response.status().is_success(),
+        Err(_) => false,
+    }
+}
+
+pub(crate) async fn graphql_rows(
+    access: &ConfigAccess,
+    collection_name: &str,
+    query: &str,
+) -> Result<Vec<Value>> {
+    let response = access.execute(query).await?;
+    Ok(response
+        .get("data")
+        .and_then(|data| data.get(collection_name))
+        .and_then(Value::as_array)
+        .cloned()
+        .unwrap_or_default())
+}
+
+pub(crate) async fn graphql_rows_or_empty_if_collection_missing(
+    access: &ConfigAccess,
+    collection_name: &str,
+    query: &str,
+) -> Result<Vec<Value>> {
+    match graphql_rows(access, collection_name, query).await {
+        Ok(rows) => Ok(rows),
+        Err(error) if is_collection_missing_error(collection_name, &error) => Ok(Vec::new()),
+        Err(error) => Err(error),
+    }
+}
+
+pub(crate) fn is_collection_missing_error(collection_name: &str, error: &anyhow::Error) -> bool {
+    let message = error.to_string();
+    message.contains(collection_name)
+        && (message.contains("collection not found") || message.contains("Cannot query field"))
+}
+
+pub(crate) fn graphql_string_list_literal(values: &[String]) -> String {
+    format!(
+        "[{}]",
+        values
+            .iter()
+            .map(|value| format!(r#""{}""#, escape_graphql_string(value)))
+            .collect::<Vec<_>>()
+            .join(", ")
+    )
+}
+
+pub(crate) fn graphql_input_literal(value: &Value) -> Result<String> {
+    match value {
+        Value::Null => Ok("null".to_string()),
+        Value::Bool(value) => Ok(value.to_string()),
+        Value::Number(value) => Ok(value.to_string()),
+        Value::String(value) => Ok(graphql_string_literal(value)),
+        Value::Array(values) => {
+            let rendered = values
+                .iter()
+                .map(graphql_input_literal)
+                .collect::<Result<Vec<_>>>()?;
+            Ok(format!("[{}]", rendered.join(", ")))
+        }
+        Value::Object(map) => {
+            let rendered = map
+                .iter()
+                .map(|(key, value)| Ok(format!("{key}: {}", graphql_input_literal(value)?)))
+                .collect::<Result<Vec<_>>>()?;
+            Ok(format!("{{ {} }}", rendered.join(", ")))
+        }
+    }
+}
+
+fn graphql_string_literal(value: &str) -> String {
+    format!(r#""{}""#, escape_graphql_string(value))
+}
+
+pub(crate) async fn post_graphql(graphql: &str, query: &str) -> Result<serde_json::Value> {
+    let client = reqwest::Client::new();
+    let response = client
+        .post(graphql)
+        .json(&json!({ "query": query }))
+        .send()
+        .await
+        .map_err(|error| {
+            anyhow::anyhow!(
+                "failed to post GraphQL to {graphql}: {error}\n{}",
+                graphql_diagnostic_hint(graphql)
+            )
+        })?;
+    let value: serde_json::Value = response.json().await.map_err(|error| {
+        anyhow::anyhow!(
+            "failed to decode GraphQL response from {graphql}: {error}\n{}",
+            graphql_diagnostic_hint(graphql)
+        )
+    })?;
+    if let Some(errors) = value.get("errors") {
+        anyhow::bail!(
+            "graphql returned errors from {graphql}: {errors}\n{}",
+            graphql_diagnostic_hint(graphql)
+        );
+    }
+    Ok(value)
+}
+
+pub(crate) fn extract_mutation_doc_id(response: &Value, collection_name: &str) -> Result<String> {
+    let data = response
+        .get("data")
+        .ok_or_else(|| anyhow::anyhow!("graphql response missing data: {response}"))?;
+    for field_name in [
+        format!("upsert_{collection_name}"),
+        format!("update_{collection_name}"),
+        format!("create_{collection_name}"),
+        format!("add_{collection_name}"),
+    ] {
+        if let Some(doc_id) = data
+            .get(&field_name)
+            .and_then(|value| value.get("_docID"))
+            .and_then(Value::as_str)
+        {
+            return Ok(doc_id.to_string());
+        }
+        if let Some(doc_id) = data
+            .get(&field_name)
+            .and_then(Value::as_array)
+            .and_then(|rows| rows.first())
+            .and_then(|row| row.get("_docID"))
+            .and_then(Value::as_str)
+        {
+            return Ok(doc_id.to_string());
+        }
+    }
+    anyhow::bail!("graphql mutation returned no _docID for {collection_name}: {response}");
+}
+
+pub(crate) fn nullable_string_field(name: &str, value: Option<&str>) -> String {
+    match value.map(str::trim).filter(|value| !value.is_empty()) {
+        Some(value) => format!(r#"{name}: "{}""#, escape_graphql_string(value)),
+        None => format!("{name}: null"),
+    }
+}
+
+pub(crate) fn graphql_bool_literal(value: bool) -> &'static str {
+    if value {
+        "true"
+    } else {
+        "false"
+    }
+}
+
+pub(crate) fn normalize_optional_rfc3339(value: Option<&str>) -> Result<Option<String>> {
+    match value.map(str::trim).filter(|value| !value.is_empty()) {
+        Some(raw) => {
+            let parsed = chrono::DateTime::parse_from_rfc3339(raw)
+                .with_context(|| format!("parsing RFC3339 timestamp {raw}"))?;
+            Ok(Some(
+                parsed
+                    .with_timezone(&chrono::Utc)
+                    .to_rfc3339_opts(chrono::SecondsFormat::Secs, true),
+            ))
+        }
+        None => Ok(None),
+    }
+}
+
+pub(crate) fn optional_i64_field(name: &str, value: Option<i64>) -> Option<String> {
+    value.map(|value| format!("{name}: {value}"))
+}
+
+pub(crate) fn optional_f64_field(name: &str, value: Option<f64>) -> Option<String> {
+    value.map(|value| format!("{name}: {value}"))
+}
+
+pub(crate) fn optional_bool_field(name: &str, value: Option<bool>) -> Option<String> {
+    value.map(|value| format!("{name}: {}", graphql_bool_literal(value)))
+}
+
+pub(crate) fn optional_string_field(name: &str, value: Option<&str>) -> Option<String> {
+    value
+        .map(str::trim)
+        .filter(|value| !value.is_empty())
+        .map(|value| format!(r#"{name}: "{}""#, escape_graphql_string(value)))
+}
+
+pub(crate) fn string_list_field(name: &str, values: &[String]) -> Option<String> {
+    Some(format!(
+        "{name}: [{}]",
+        values
+            .iter()
+            .map(|value| format!(r#""{}""#, escape_graphql_string(value)))
+            .collect::<Vec<_>>()
+            .join(", ")
+    ))
+}
+
+fn is_probably_local_graphql_endpoint(graphql: &str) -> bool {
+    let graphql = graphql.trim();
+    graphql.contains("127.0.0.1") || graphql.contains("localhost")
+}
+
+pub(crate) fn graphql_diagnostic_hint(graphql: &str) -> String {
+    if is_probably_local_graphql_endpoint(graphql) {
+        "Next:\n  1. If this home is not initialized, run `defra-agent init`\n  2. Start the runtime with `defra-agent server`\n  3. Inspect it with `defra-agent status`".to_string()
+    } else {
+        format!(
+            "Next:\n  1. Verify the GraphQL endpoint {graphql}\n  2. Retry with `--graphql {graphql}` or point the command at the correct runtime"
+        )
+    }
+}
+
+pub(crate) fn first_graphql_row<'a>(response: &'a Value, collection_name: &str) -> Result<&'a Value> {
+    response
+        .get("data")
+        .and_then(|data| data.get(collection_name))
+        .and_then(Value::as_array)
+        .and_then(|rows| rows.first())
+        .ok_or_else(|| anyhow::anyhow!("graphql returned no rows for {collection_name}"))
+}
